requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
codemelli==0.1.2
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.5
django-appconf==1.0.5
django-cleanup==6.0.0
django-crum==0.7.9
django-filter==21.1
django-ipware==4.0.2
django-jalali==6.0.0
django-jalali-date==1.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
fpdf==1.7.2
idna==3.3
jalali==1.2
jdatetime==4.1.0
kombu==5.2.4
mysqlclient==2.1.0
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
python-magic==0.4.27
pytz==2022.1
redis==4.3.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
